requirements.txt
argparse
httplib2>=0.7.0
jinja2
lxml
zope.dottedname
